Renamed window file so we can add more stuff in the same file and it makes sense
