+++ main.py
@@ -1,4 +1,4 @@
-from window import Window
+from graphics import Window
 
 def main():
     window = Window(800, 600)
